feat(maintenance)!: ajout d'un contexte de maintenance
chore(architecture): factorisation de types de configuration
