+++ src/maintenance/start.ts
@@ -1,10 +1,10 @@
 import * as http from "http";
 import { Client } from "minio";
 
-import { ConfigurationFactory } from "@configuration/configuration";
+import { ConfigurationFactory } from "@maintenance/configuration";
+import { Setup } from "@maintenance/setup";
 import { LoggerFactory } from "@shared/configuration/logger";
 import { MinioAdminStorageRepository } from "@shared/infrastructure/gateway/repository/minio-admin-storage.repository";
-import { Setup } from "@configuration/setup";
 
 const configuration = ConfigurationFactory.create();
 const loggerFactory = new LoggerFactory(
