+++ src/shared/configuration/index.ts
@@ -1,0 +1,11 @@
+export enum Environment {
+	DEVELOPMENT = "development",
+	PRODUCTION = "production",
+	QUALIFICATION = "qualification"
+}
+
+export type SentryConfiguration = {
+	DSN: string
+	PROJECT: string
+	RELEASE: string
+}
